refactor: controller to service
contorller에서 만들던 dto를 service로 이동

[#19]
