+++ packages/backend/src/problems/problems.module.ts
@@ -3,7 +3,7 @@ import { ProblemsController } from './problems.controller';
 import { ProblemsService } from './problems.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Problem } from './entity/problem.entity';
-import {Category} from "./entity/category.entity";
+import { Category } from './entity/category.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Problem, Category])],
