+++ packages/backend/src/problems/problems.service.ts
@@ -4,6 +4,7 @@ import { Problem } from './entity/problem.entity';
 import { Repository } from 'typeorm';
 import { Category } from './entity/category.entity';
 import { CategoryProblemsDto } from './dto/problems.dto';
+import { ProblemDto } from './dto/problem.dto';
 
 @Injectable()
 export class ProblemsService {
@@ -14,7 +15,7 @@ export class ProblemsService {
     private categoryRepository: Repository<Category>,
   ) {}
 
-  async getProblemById(id: number): Promise<Problem> {
+  async getProblemById(id: number): Promise<ProblemDto> {
     const problem = await this.problemRepository.findOne({
       where: { id },
       relations: ['keywords', 'category'],
@@ -24,7 +25,15 @@ export class ProblemsService {
       throw new NotFoundException(`Problem${id} not found`);
     }
 
-    return problem;
+    const problemDto: ProblemDto = {
+      id: problem.id,
+      title: problem.title,
+      description: problem.description,
+      keywords: problem.keywords.map((keyword) => keyword.keyword),
+      category: problem.category.name,
+    };
+
+    return problemDto;
   }
 
   async findAllProblemsGroupedByCategory(): Promise<CategoryProblemsDto[]> {
